requirements.txt
piexif
shortuuid
pypiwin32
Pillow
pillow-heif
MoviePy